% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sideway.R
\name{sideway}
\alias{sideway}
\alias{sideway.default}
\alias{sideway.data.frame}
\title{Print Sideways}
\usage{
sideway(x, ...)

\method{sideway}{default}(x, sep = ": ", colour = list(col_cyan, col_magenta), ...)

\method{sideway}{data.frame}(x, ...)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector},
or a \link[base]{list} of \link[base]{character} object.
Input \code{x} must be named}

\item{...}{additional parameters for future use}

\item{sep}{\link[base]{character} scalar, see \link[base]{paste}}

\item{colour}{\link[base]{list} of color functions from package \CRANpkg{cli}, default
\code{list(col_cyan, col_magenta)}}
}
\value{
Function \code{\link[=sideway]{sideway()}} does not have a return value.
}
\description{
Print Sideways
}
\examples{
c(a = 'a1', bc = '2\n3') |> sideway()
list(a = '1\n2', b = character(), cd = '3\n4', efg = '5\n6\n7') |> sideway()
c(a = '1\n2') |> sideway()
# data.frame
Formaldehyde |> sideway() # no rownames
USArrests |> head() |> sideway() # with rownames
}
\keyword{internal}
